// src/generated/graphql.tsx
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: any;
  /** An ISO-8601 encoded UTC date string. */
  DateTime: any;
  /** A string containing HTML code. */
  HTML: any;
  /** A Git object ID. */
  GitObjectID: any;
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: any;
  /** Git SSH string */
  GitSSHRemote: any;
  /** An ISO-8601 encoded date string. */
  Date: any;
  /** An ISO-8601 encoded UTC date string with millisecond precison. */
  PreciseDateTime: any;
  /** A valid x509 certificate string */
  X509Certificate: any;
};
// ...
export type Blob = GitObject &
  Node & {
    __typename?: "Blob";
    /** An abbreviated version of the Git object ID */
    abbreviatedOid: Scalars["String"];
    /** Byte size of Blob object */
    byteSize: Scalars["Int"];
    /** The HTTP path for this Git object */
    commitResourcePath: Scalars["URI"];
    /** The HTTP URL for this Git object */
    commitUrl: Scalars["URI"];
    id: Scalars["ID"];
    /** Indicates whether the Blob is binary or text */
    isBinary: Scalars["Boolean"];
    /** Indicates whether the contents is truncated */
    isTruncated: Scalars["Boolean"];
    /** The Git object ID */
    oid: Scalars["GitObjectID"];
    /** The Repository the Git object belongs to */
    repository: Repository;
    /** UTF8 text data or null if the Blob is binary */
    text?: Maybe<Scalars["String"]>;
  };
// ...
export type GitObject = {
  __typename?: "GitObject";
  /** An abbreviated version of the Git object ID */
  abbreviatedOid: Scalars["String"];
  /** The HTTP path for this Git object */
  commitResourcePath: Scalars["URI"];
  /** The HTTP URL for this Git object */
  commitUrl: Scalars["URI"];
  id: Scalars["ID"];
  /** The Git object ID */
  oid: Scalars["GitObjectID"];
  /** The Repository the Git object belongs to */
  repository: Repository;
};
// ...
export type Language = Node & {
  __typename?: "Language";
  /** The color defined for the current language. */
  color?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  /** The name of the current language. */
  name: Scalars["String"];
};
// ...
export type Node = {
  __typename?: "Node";
  /** ID of the object. */
  id: Scalars["ID"];
};
// ...
export type PageInfo = {
  __typename?: "PageInfo";
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars["String"]>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars["Boolean"];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars["Boolean"];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars["String"]>;
};
// ...
export type RepositoryConnection = {
  __typename?: "RepositoryConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<RepositoryEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<Repository>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
  /** The total size in kilobytes of all repositories in the connection. */
  totalDiskUsage: Scalars["Int"];
};
// ...
export type RepositoryEdge = {
  __typename?: "RepositoryEdge";
  /** A cursor for use in pagination. */
  cursor: Scalars["String"];
  /** The item at the end of the edge. */
  node?: Maybe<Repository>;
};
// ...
export type RepositoryOwner = {
  __typename?: "RepositoryOwner";
  /** A URL pointing to the owner's public avatar. */
  avatarUrl: Scalars["URI"];
  id: Scalars["ID"];
  /** The username used to login. */
  login: Scalars["String"];
  /** A list of repositories this user has pinned to their profile */
  pinnedRepositories: RepositoryConnection;
  /** A list of repositories that the user owns. */
  repositories: RepositoryConnection;
  /** Find Repository. */
  repository?: Maybe<Repository>;
  /** The HTTP URL for the owner. */
  resourcePath: Scalars["URI"];
  /** The HTTP URL for the owner. */
  url: Scalars["URI"];
};
// ...
export type StargazerConnection = {
  __typename?: "StargazerConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<StargazerEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<User>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};

/** Represents a user that's starred a repository. */
export type StargazerEdge = {
  __typename?: "StargazerEdge";
  /** A cursor for use in pagination. */
  cursor: Scalars["String"];
  node: User;
  /** Identifies when the item was starred. */
  starredAt: Scalars["DateTime"];
};
// ...
export interface IntrospectionResultData {
  __schema: {
    types: {
      kind: string;
      name: string;
      possibleTypes: {
        name: string;
      }[];
    }[];
  };
}